Added DateRangePicker; js_only.
