+++ assets/prototype/domain/eventEditor/datesList/dateCard/style.css
@@ -1,0 +1,3 @@
+.bp3-popover-content .ee-date-range-picker {
+	padding: 1rem;
+}
