+++ assets/prototype/domain/shared/dateRangeInput/momentDate.js
@@ -9,7 +9,7 @@ const FORMAT_TIME = 'dddd, LL LT';
 export const MomentDate = ({ date, withTime = false, format = withTime ? FORMAT_TIME : FORMAT }) => {
 	const m = moment(date);
 	if (m.isValid()) {
-		return <Tag intent={Intent.PRIMARY}>{m.format(format)}</Tag>;
+		return <Tag intent={Intent.SUCCESS}>{m.format(format)}</Tag>;
 	} else {
 		return <Tag minimal={true}>no date</Tag>;
 	}
